Move CopyFilter to textsub.py
